优化歌词错误提示 优化id3的tpos显示，

+++ file.py
@@ -1,5 +1,6 @@
 import os
 
+
 def rename(fileName, track_no, artist, title, discnumber, totaldiscs):
     # 转换成简体文件名， 重命名文件
     (path, oldName) = os.path.split(fileName)
@@ -16,9 +17,11 @@ def rename(fileName, track_no, artist, title, discnumber, totaldiscs):
     # 寻找同级目录下的lrc歌词文件并重命名
     (tmp_path, ext) = os.path.splitext(fileName)
     lrc_file = tmp_path + '.lrc'
-    new_lrc_path = os.path.join(path, newfile + '.lrc')
-    try:
-        os.rename(lrc_file, new_lrc_path)
-    except OSError as e:
-        print(f'rename lrc error: {e.errno}, Error text: {e.strerror}')
-    print(newpath + ' done')
+    if os.path.exists():
+        new_lrc_path = os.path.join(path, newfile + '.lrc')
+        try:
+            os.rename(lrc_file, new_lrc_path)
+        except OSError as e:
+            print(f'rename lrc error: {e.errno}, Error text: {e.strerror}')
+
+    print(newfile + ext + ' done')
